Long titles (#30)
* added mockdata

* added mockdata

* removed whitespace-nowrap

* added long title

* "ok"Merge branch 'main' into longTitles

Co-authored-by: Isac Zetterström <IsacZetterstrom@users.noreply.github.com>
added function that gives hyphen and new lins if title is longer or 21 char. Works on small screen now, but bug on tablet and desktop still.

* Co-authored-by: Isac Zetterström

added new mockData to try out our function

* Co-Authored-by:IsacZetterxstrom:
added breakwords and hyphens-auto to break a title in to a new line if its to long.